Moved remote index build client to a separate module
