+++ remote-index-build-client/NOTICE.txt
@@ -1,0 +1,2 @@
+OpenSearch k-NN
+Copyright OpenSearch Contributors
