+++ remote-index-build-client/build.gradle
@@ -1,0 +1,70 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+plugins {
+    id 'java'
+    id 'java-library'
+    id 'jacoco'
+    id "io.freefair.lombok"
+    id 'com.diffplug.spotless' version '6.25.0'
+    id 'opensearch.build'
+}
+
+repositories {
+    mavenLocal()
+    maven { url = uri("https://aws.oss.sonatype.org/content/repositories/snapshots") }
+    mavenCentral()
+    maven { url = uri("https://plugins.gradle.org/m2/") }
+}
+
+description = "Remote Index Build Client for OpenSearch"
+
+test {
+    systemProperty "jdk.attach.allowAttachSelf", true
+}
+
+compileJava {
+    options.compilerArgs = ["-processor", 'lombok.launch.AnnotationProcessorHider$AnnotationProcessor']
+}
+compileTestJava {
+    options.compilerArgs = ["-processor", 'lombok.launch.AnnotationProcessorHider$AnnotationProcessor']
+}
+
+dependencies {
+    compileOnly group: 'org.opensearch', name: 'opensearch', version: "${opensearch_version}"
+    api "commons-lang:commons-lang:2.6"
+    api "org.apache.httpcomponents.client5:httpclient5:${versions.httpclient5}"
+    api "org.apache.httpcomponents.core5:httpcore5:${versions.httpcore5}"
+    api "org.apache.httpcomponents.core5:httpcore5-h2:${versions.httpcore5}"
+    testImplementation "org.opensearch.test:framework:${opensearch_version}"
+    testImplementation group: 'net.bytebuddy', name: 'byte-buddy', version: "${versions.bytebuddy}"
+    testImplementation group: 'org.objenesis', name: 'objenesis', version: "${versions.objenesis}"
+    testImplementation group: 'net.bytebuddy', name: 'byte-buddy-agent', version: "${versions.bytebuddy}"
+}
+
+check.dependsOn spotlessCheck
+
+publishing {
+    publications {
+        pluginZip(MavenPublication) {
+            pom {
+                name = "opensearch-knn-remote-index-build-client"
+                description = 'Client for connecting to remote index build endpoint'
+                licenses {
+                    license {
+                        name = "The Apache License, Version 2.0"
+                        url = "http://www.apache.org/licenses/LICENSE-2.0.txt"
+                    }
+                }
+                developers {
+                    developer {
+                        name = "OpenSearch"
+                        url = "https://github.com/opensearch-project/k-NN/remote-index-build-client"
+                    }
+                }
+            }
+        }
+    }
+}
