+++ remote-index-build-client/src/main/java/org/opensearch/remoteindexbuild/model/RemoteBuildStatusRequest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.knn.index.remote;
+package org.opensearch.remoteindexbuild.model;
 
 import lombok.Builder;
 import lombok.Getter;
