+++ remote-index-build-client/src/test/java/org/opensearch/remoteindexbuild/TestConstants.java
@@ -1,0 +1,24 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package org.opensearch.remoteindexbuild;
+
+public class TestConstants {
+    public static final String TEST_BUCKET = "test-bucket";
+    public static final String TEST_CLUSTER = "test-cluster";
+    public static final String MOCK_JOB_ID_RESPONSE = "{\"job_id\": \"job-1739930402\"}";
+    public static final String MOCK_JOB_ID = "job-1739930402";
+    public static final String MOCK_BLOB_NAME = "blob";
+    public static final String MOCK_ENDPOINT = "https://mock-build-service.com";
+    public static final String USERNAME = "username";
+    public static final String PASSWORD = "password";
+    public static final String L2_SPACE_TYPE = "l2";
+    public static final String HNSW_ALGORITHM = "hnsw";
+    public static final String CONTAINER = "container";
+    public static final String FLOAT = "float";
+    public static final String FAISS = "faiss";
+    public static final String UNKNOWN_FIELD = "unknown_field";
+    public static final String UNKNOWN_VALUE = "value";
+}
