+++ src/main/java/org/opensearch/knn/index/remote/RemoteIndexWaiterFactory.java
@@ -5,6 +5,8 @@
 
 package org.opensearch.knn.index.remote;
 
+import org.opensearch.remoteindexbuild.client.RemoteIndexClient;
+
 public class RemoteIndexWaiterFactory {
 
     /**
