+++ src/test/java/org/opensearch/knn/index/codec/nativeindex/remote/RemoteIndexBuildStrategyTests.java
@@ -13,9 +13,11 @@ import org.opensearch.common.settings.ClusterSettings;
 import org.opensearch.core.common.unit.ByteSizeValue;
 import org.opensearch.core.index.Index;
 import org.opensearch.index.IndexSettings;
+import org.opensearch.knn.common.KNNConstants;
 import org.opensearch.knn.index.KNNSettings;
-import org.opensearch.knn.index.codec.nativeindex.model.BuildIndexParams;
-import org.opensearch.knn.index.remote.RemoteIndexHTTPClientTests;
+import org.opensearch.knn.index.VectorDataType;
+import org.opensearch.knn.index.engine.KNNEngineTests;
+import org.opensearch.remoteindexbuild.model.RemoteBuildRequest;
 import org.opensearch.repositories.RepositoriesService;
 import org.opensearch.repositories.RepositoryMissingException;
 import org.opensearch.repositories.blobstore.BlobStoreRepository;
@@ -28,13 +30,17 @@ import static org.mockito.Mockito.when;
 import static org.opensearch.knn.index.KNNSettings.KNN_INDEX_REMOTE_VECTOR_BUILD_SETTING;
 import static org.opensearch.knn.index.KNNSettings.KNN_INDEX_REMOTE_VECTOR_BUILD_THRESHOLD_SETTING;
 import static org.opensearch.knn.index.KNNSettings.KNN_REMOTE_VECTOR_REPO_SETTING;
-import static org.opensearch.knn.index.remote.RemoteBuildRequestTests.MOCK_BASE_PATH;
-import static org.opensearch.knn.index.remote.RemoteBuildRequestTests.MOCK_SEGMENT_STATE;
-import static org.opensearch.knn.index.remote.RemoteBuildRequestTests.MOCK_UUID;
-import static org.opensearch.knn.index.remote.RemoteBuildRequestTests.VECTORS_PATH;
+import static org.opensearch.remoteindexbuild.constants.KNNRemoteConstants.DOC_ID_FILE_EXTENSION;
+import static org.opensearch.remoteindexbuild.constants.KNNRemoteConstants.S3;
+import static org.opensearch.remoteindexbuild.constants.KNNRemoteConstants.VECTOR_BLOB_FILE_EXTENSION;
 
 public class RemoteIndexBuildStrategyTests extends RemoteIndexBuildTests {
     private static final String TEST_INDEX = "test-index";
+    public static final String MOCK_BASE_PATH = "vectors/1_1_25";
+    public static final String MOCK_UUID = "SIRKos4rOWlMA62PX2p75m";
+    public static final String VECTORS_PATH = "_vectors";
+    public static final String MOCK_FULL_PATH = "vectors/1_1_25/SIRKos4rOWlMA62PX2p75m_vectors/SIRKos4rOWlMA62PX2p75m_target_field__3l";
+    public static final String MOCK_SEGMENT_STATE = "_3l";
 
     /**
      * Test that we fallback to the fallback NativeIndexBuildStrategy when an exception is thrown
@@ -99,15 +105,14 @@ public class RemoteIndexBuildStrategyTests extends RemoteIndexBuildTests {
         BlobPath baseBlobPath = new BlobPath().add(MOCK_BASE_PATH);
         when(repository.basePath()).thenReturn(baseBlobPath);
 
-        IndexSettings indexSettings = RemoteIndexHTTPClientTests.createTestIndexSettings();
+        IndexSettings indexSettings = createTestIndexSettings();
         BlobPath blobPath = repository.basePath().add(indexSettings.getUUID() + VECTORS_PATH);
 
         BlobContainer blobContainer = mock(BlobContainer.class);
         when(mockBlobStore.blobContainer(blobPath)).thenReturn(blobContainer);
         when(blobContainer.path()).thenReturn(blobPath);
 
-        BuildIndexParams indexInfo = RemoteIndexHTTPClientTests.createTestBuildIndexParams();
-        String blobName = MOCK_UUID + "_" + indexInfo.getFieldName() + "_" + MOCK_SEGMENT_STATE;
+        String blobName = MOCK_UUID + "_" + buildIndexParams.getFieldName() + "_" + MOCK_SEGMENT_STATE;
 
         // example: VectorRepositoryAccessor vectorAccessor = new DefaultVectorRepositoryAccessor(blobContainer);
         // vectorAccessor.writeToRepository(blobName...)
@@ -115,5 +120,24 @@ public class RemoteIndexBuildStrategyTests extends RemoteIndexBuildTests {
 
         String requestPath = blobPath.buildAsString() + blobName;
         assertEquals(accessorPath + blobName, requestPath);
+    }
+
+    public void testBuildRequest() throws IOException {
+        RemoteBuildRequest request = RemoteIndexBuildStrategy.buildRemoteBuildRequest(
+            createTestIndexSettings(),
+            buildIndexParams,
+            createTestRepositoryMetadata(),
+            MOCK_FULL_PATH,
+            KNNEngineTests.createMockMethodContext()
+        );
+        assertEquals(S3, request.getRepositoryType());
+        assertEquals(TEST_BUCKET, request.getContainerName());
+        assertEquals(KNNConstants.FAISS_NAME, request.getEngine());
+        assertEquals(VectorDataType.FLOAT.getValue(), request.getVectorDataType());
+        assertEquals(MOCK_FULL_PATH + VECTOR_BLOB_FILE_EXTENSION, request.getVectorPath());
+        assertEquals(MOCK_FULL_PATH + DOC_ID_FILE_EXTENSION, request.getDocIdPath());
+        assertEquals(TEST_CLUSTER, request.getTenantId());
+        assertEquals(3, request.getDocCount());
+        assertEquals(2, request.getDimension());
     }
 }
